[놋북] 새로그린 ERD 따라서 프로젝트 갈아 엎는중

+++ src/main/java/me/sailer/my_atelier/dto/UpdateProductRequest.java
@@ -20,6 +20,6 @@ public class UpdateProductRequest {
     private int remaining;
     private boolean isDiscounted;
     private int discountedPrice;
-    private long categoryId;
+    private byte categoryId;
 
 }
